loader script
Chrome extensions don't yet support modules for content scripts, so instead it's necessary to use a small standard script as the entry point and then use that to add the module to the page. JSDOM doesn't play well with this approach, so the build can be used to compile a more straightforward script which is used for testing purposes only and then immediately deleted by the lifecycle hooks.