store all code in content script
Splitting code into separate ES6 modules is nice for architecture and
testing, but it complicates initialization by requiring an additional
loader script and even in the best case scenario would leave the DOM
manipulation in separate files which would be subject to the page's
Content Security Policy. Allowing the page to block the extension
undermines the whole point, so it's a better tradeoff to smash the
architecture together into a single file which will always load, even on
hostile pages.